transform: ...the hard-coding site, mock-up to react rendered page
The site is designed to show the main page of the bookshelf,
having three shelfs with books to store during the user reading them. There is
a search page where the user finds books from a library too. To transform html to jsx
in favor of react implementation a new component of main page implemented, a search page too.
In the hard coding site the book html code was in many cases, so a book of these taken in order to
implement react component that renders book template. Any hard coding of books removed.
Html code of shelfs kept as it was, to make the three selfs of the bookshelf application.

React is to have many components structured like a tree, so future breaking the app
in more parts will make it conform the react philosophy.