feat: implement functionality to application's components
When the main component of the application is ready, then
use the get all book method of the backend to load the book info to the app, so to be ready. Make the main page component to display the chosen books at the right shelf of the bookshelf, this implemented using obj.filter().map() chain of functions. Make the book component to display thumbnail of the cover of the book and title and authors info, by using props to pass data to the html elements.
Implement the move shelf function of the app, so that the user to have the ability of move book to another shelf, according the reading face of it. To do that a backend method used that updates the data of the book.

Some hack used in order to pass props to the component <div> that of {condition ? true : false} and needs extra research
 to be done to examine why these props occured error of object not defined.