feat: implement search functionality on search component
Search functionality is triggered on queries the user inputs,
from the first character input the results of the searched books is displayed. If the query returns books with no book cover thumbnail, it returns an empty thumbnail. A way to catch the error on resulting book search is done.

A placeholder for the book with no cover image is good to be implemented.