feat: implement application routing  and rearch result improve
Wrap app element in browser router. Give route at main and search page. Give the unassigned books the out of the bookshelf sign. Pass with the book component the current shelf to the props. Change the ancor element with link component, so that routing to be assigned to the URL.